test(people): the open-duplicates badge obeys the pair's both-sides rule
CountOpenDedupeCandidates already applies it — the production side of this
ticket landed with the digest's per-reader seam. What had no test was the count,
and the list's did not cover it: the two assemble their clause separately, so a
slip in one spelling does not show in the other.

The number is what the daily digest reports and what the attention card badges,
and it is an invitation to open a queue. Taken without the rule it is two
defects at once: the reader is told twelve and shown three, and it moves as
records they may not open come and go — a weak existence oracle over a
colleague's capture-private contact.

Over the same arms and both slots as the list, from one builder: an arm added
for one surface that the other never ran is exactly the gap this closes. The
owner of the private half is asserted to count 1 first, so the colleague's 0 is
the rule refusing rather than an empty queue answering for free.

Mutation-checked: dropping the visibility clause from the count fails all four
cases.

The half-visible cases move to their own file with the fixtures they share. The
queue file crossed the 1000-line test ceiling, and the rule asked of three
statements is the concept that comes out whole.
